Update RewardEmpireDestroyer.xml
Edits the throne room and bedroom to meet the minimum requirements  for the proper title that you would receive this ship at.

However, sometimes impressiveness will not be met in the throne room due to the RNG on quality (the baseline I used was all normal to reach 161 impressiveness). The player can simply rebuild any awful or poor furniture instead of having to redesign the whole room.