RewardEmpireDestroyerFixForUnstable
Modifies the ship throne room and bedroom to meet the minimum requirements for the players title that they would receive the ship at.

The throne room may be undignified due to RNG on the quality of the furniture (normal quality was used for the baseline). This can be fixed by the player rebuilding it, which is easier than trying to re-design the room to fit the requirements.